requirements.txt
Flask
flask-cors
groq
markdown
PyPDF2
python-dotenv
